refactor(core): extract MiddlewareSlot enum + SLOTS_PER_ROLE to decepticon-core (Phase 1.C)
Per docs/superpowers/specs/2026-05-23-core-framework-sdk-split-design.md
§9.1 — split decepticon/agents/middleware_slots.py: the pure-data
declarations (the enum, the per-role applicability mapping, the
safety-critical slot set) move into the contract layer; the
langchain/deepagents-bound default factories stay in the framework.

New core module:

  packages/decepticon-core/src/decepticon_core/contracts/__init__.py
  packages/decepticon-core/src/decepticon_core/contracts/slots.py

contracts/slots.py carries: MiddlewareSlot StrEnum, SAFETY_CRITICAL_SLOTS,
_TAIL_SLOTS / _BASE_SLOTS / _BASH_AGENT_SLOTS helpers, and the full
SLOTS_PER_ROLE mapping for the 16 OSS roles.

Framework-side middleware_slots.py becomes ~half its previous size:
imports MiddlewareSlot / SAFETY_CRITICAL_SLOTS / SLOTS_PER_ROLE from
core and re-exports them via ``__all__`` (so existing
``from decepticon.agents.middleware_slots import SLOTS_PER_ROLE`` calls
in build.py and the test tree keep working). Local responsibilities
remaining there: DEFAULT_SLOT_FACTORIES, the _make_* factory helpers
(all langchain / deepagents-bound), and skills_sources_for().

Phase 1.C gate:
  uv run ruff check .            ->  All checks passed!
  uv run ruff format --check .   ->  green
  uv run basedpyright --level error ->  0 errors, 0 warnings, 0 notes
  uv sync                        ->  119 packages OK
  enum identity check            ->  MS_core is MS_fw (True)
  factories count                ->  11 (all roles preserved)
  roles known                    ->  16 (full SLOTS_PER_ROLE)
  skills_sources_for("recon")    ->  ['/skills/standard/recon/', '/skills/shared/']

Remaining Phase 1 work:
  1.D: add Protocols (BackendProtocol, MiddlewareProtocol, ...)
  1.E: add new registries (RoleRegistry, SkillSourceRegistry, PluginRegistry.introspect)
  1.F: add new contributions + ruff banned-imports rule for core
